Something like this?
So, ideally, we could add an `embed` bool and use that to signify if the quote should embed its source (where valid). Then, we don't have to do "attribute only" in `format_quote_text` and can have both the quote and an embed, such as with the Pale Blue Dot.

Also, for the sake of finding future bugs, perhaps we can add some logging to `quote_compliant` so it tells us in which step it failed, so when a ill-formed quote is added it can be tracked down faster.